4.0.3 - Nueva tarea updateCheck para comprobar actualizaciones del Framework

+++ ofw/task/update.php
@@ -9,31 +9,45 @@ class updateTask{
   function doUpdate($repo_version){
     global $c;
     $updates = json_decode( file_get_contents($this->repo_url.'ofw/base/updates.json'), true );
-    echo "Actualizando a versión: ".$repo_version.":\n";
-    echo "  ".$updates[$repo_version]['message']."\n\n";
+    echo "  ".$updates[$repo_version]['message']."\n";
+    echo "==============================================================================================================\n\n";
     $backups = [];
     $result = true;
 
-    foreach ($updates[$repo_version]['files'] as $file){
-      $file_url = $this->repo_url.$file;
-      echo "Descargando ".$file_url."\n";
-      $file_content = file_get_contents($file_url);
-
-      $local_file = $c->getDir('base').$file;
-      if (file_exists($local_file)){
-        echo "  El archivo ya existe, creando copia de seguridad\n";
-        $backup_file = $local_file.'_backup';
-        rename($local_file, $backup_file);
-        array_push($backups, ['new_file'=>$local_file, 'backup'=>$backup_file]);
+    if (count($updates[$repo_version]['deletes'])>0){
+      foreach ($updates[$repo_version]['deletes'] as $delete){
+        $local_delete = $c->getDir('base').$delete;
+        if (file_exists($local_delete)){
+          echo " El archivo \"".$local_delete."\" será eliminado.\n";
+          $backup_file = $local_delete.'_backup';
+          rename($local_delete, $backup_file);
+          array_push($backups, ['new_file'=>$local_delete, 'backup'=>$backup_file]);
+        }
       }
-      else{
-        echo "Creando nuevo archivo\n";
-      }
+      echo "\n";
+    }
+    if (count($updates[$repo_version]['files'])>0){
+      foreach ($updates[$repo_version]['files'] as $file){
+        $file_url = $this->repo_url.$file;
+        echo "Descargando \"".$file_url."\"\n";
+        $file_content = file_get_contents($file_url);
 
-      $result_file = file_put_contents($local_file, $file_content);
-      if ($result_file===false){
-        $result = false;
-        break;
+        $local_file = $c->getDir('base').$file;
+        if (file_exists($local_file)){
+          echo "  El archivo ya existe, creando copia de seguridad.\n";
+          $backup_file = $local_file.'_backup';
+          rename($local_file, $backup_file);
+          array_push($backups, ['new_file'=>$local_file, 'backup'=>$backup_file]);
+        }
+        else{
+          echo "Creando nuevo archivo.\n";
+        }
+
+        $result_file = file_put_contents($local_file, $file_content);
+        if ($result_file===false){
+          $result = false;
+          break;
+        }
       }
     }
 
@@ -51,7 +65,9 @@ class updateTask{
     else{
       echo "Ocurrió un error al actualizar los archivos, se procede a restaurar las copias de seguridad.\n";
       foreach ($backups as $backup){
-        unlink($backup['new_file']);
+        if (file_exists($backup['new_file'])){
+          unlink($backup['new_file']);
+        }
         rename($backup['backup'], $backup['new_file']);
       }
     }
